Move Part components to Part
