+++ frontend/src/Activity/Garmin.svelte
@@ -12,7 +12,7 @@
   } from "@sveltestrap/sveltestrap";
   import { attachments, checkStatus, handleError } from "../lib/store";
   import TZPicker from "../Widgets/TZPicker.svelte";
-  import ModalFooter from "./ModalFooter.svelte";
+  import ModalFooter from "../Widgets/ModalFooter.svelte";
   import { parts } from "../Part/part";
 
   let timezone: string | undefined;
