+++ frontend/src/Part/ChangePart.svelte
@@ -8,13 +8,13 @@
     Form,
   } from "@sveltestrap/sveltestrap";
   import { handleError, types, attachments } from "../lib/store";
-  import ModalFooter from "./ModalFooter.svelte";
+  import ModalFooter from "../Widgets/ModalFooter.svelte";
   import { AttEvent, Attachment, maxDate } from "../lib/types";
-  import NewForm from "./NewForm.svelte";
+  import NewForm from "./PartForm.svelte";
   import Dispose from "../Widgets/Dispose.svelte";
   import DateTime from "../Widgets/DateTime.svelte";
   import Switch from "../Widgets/Switch.svelte";
-  import { Part } from "../Part/part";
+  import { Part } from "./part";
 
   let atts: Attachment[];
   let last: Attachment;
