+++ frontend/src/Part/PartForm.svelte
@@ -9,7 +9,7 @@
   import DateTime from "../Widgets/DateTime.svelte";
   import { Type } from "../lib/types";
   import { createEventDispatcher } from "svelte";
-  import { Part } from "../Part/part";
+  import { Part } from "./part";
   const dispatch = createEventDispatcher();
 
   export let type: Type | undefined;
