+++ frontend/src/Service/NewService.svelte
@@ -5,7 +5,7 @@
     ModalHeader,
     Form,
   } from "@sveltestrap/sveltestrap";
-  import ModalFooter from "../Actions/ModalFooter.svelte";
+  import ModalFooter from "../Widgets/ModalFooter.svelte";
   import { Service } from "./service";
   import { Part } from "../Part/part";
   import ServiceForm from "./ServiceForm.svelte";
