+++ frontend/src/Widgets/ModalFooter.svelte
